+++ GetThePicture/Cobol/Base/Area.cs
@@ -1,0 +1,7 @@
+namespace GetThePicture.Cobol.Base;
+
+public enum Area_t
+{
+    None,
+    A, B, Free,
+}